refactor (#293)
